+++ Ticket Management System/Controllers/TicketPriorityController.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Ticket_Management_System.Contracts;
+using Ticket_Management_System.DTOs.DepartmentDTO;
+using Ticket_Management_System.DTOs.EmployeeDTO;
 using Ticket_Management_System.DTOs.TicketPriorityDTO;
 
 namespace Ticket_Management_System.Controllers
@@ -9,7 +11,7 @@ namespace Ticket_Management_System.Controllers
     /// </summary>
     [Route("api/[controller]")]
     [ApiController]
-    public class TicketPriorityController : ControllerBase
+    public class TicketPriorityController : BaseController
     {
         private readonly ITicketPriorityService _ticketPriorityService;
 
@@ -30,6 +32,8 @@ namespace Ticket_Management_System.Controllers
         [HttpPost]
         public async Task<ActionResult<TicketPriorityResponseDTO>> CreateTicketPriority(TicketPriorityRequestDTO ticketPriorityRequestDTO)
         {
+            var validation = ValidateDTO(ticketPriorityRequestDTO);
+            if (validation != null) return validation;
             var createdTicketPriority = await _ticketPriorityService.CreateTicketPriorityAsync(ticketPriorityRequestDTO);
             if (createdTicketPriority == null)
             {
@@ -47,6 +51,8 @@ namespace Ticket_Management_System.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<TicketPriorityResponseDTO>> GetTicketPriorityById(int id)
         {
+            var validationResult = ValidateId(id);
+            if (validationResult != null) return validationResult;
             var ticketPriority = await _ticketPriorityService.GetTicketPriorityByIdAsync(id);
             if (ticketPriority == null)
             {
@@ -95,6 +101,8 @@ namespace Ticket_Management_System.Controllers
         [HttpPut("{id}")]
         public async Task<ActionResult<TicketPriorityResponseDTO>> UpdateTicketPriority(int id, TicketPriorityRequestDTO ticketPriorityRequestDTO)
         {
+            var validationResult = ValidateDTOWithId(ticketPriorityRequestDTO, id);
+            if (validationResult != null) return validationResult;
             var updatedTicketPriority = await _ticketPriorityService.UpdateTicketPriorityAsync(id, ticketPriorityRequestDTO);
             if (updatedTicketPriority == null)
             {
@@ -112,6 +120,8 @@ namespace Ticket_Management_System.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult<string>> DeleteTicketPriority(int id)
         {
+            var validationResult = ValidateId(id);
+            if (validationResult != null) return validationResult;
             var result = await _ticketPriorityService.DeleteTicketPriorityAsync(id);
             if (result == null)
             {
